[testharness.js] Formalize test types
Previously, tests for the testharness.js framework were validated
according to an implicit heuristic based on the presence of a `<script>`
tag.

Introduce a declarative configuration to make this distinction explicit.
This pattern makes the validation behavior easier to understand. It also
enables future additions which further differentiate the way the two
test types are interpreted.